+++ StopRugRemoval/StopRugRemoval/docs/changelog.md
@@ -6,6 +6,10 @@
 2. Prevent grass from being placed under things when SmartBuilding's building mode is on.
 <!-- Figure out how to replace the reference to coffee in the night market dialogue?-->
 
+#### Version 1.0.8 - alpha
+Compiled for SMAPI 3.14.0
+* Removes damage to fruit trees from hoes and scythes.
+
 #### Version 1.0.7
 
 * Just-in-case compatability adjustments for Smart Building - disables this mod's ability to place grass under things whenever Smart Building's build mode is on, which IMO makes it easier to place things anyways....